+++ benchmarks/benchjs/multiop.js
@@ -19,11 +19,11 @@ const utils = require('../../lib/utils.js');
  * These are all the benchmark tests for mixed data operations
  */
 
-mongoose.connect('mongodb://localhost/mongoose-bench', function (err) {
+mongoose.connect('mongodb://127.0.0.1/mongoose-bench', function (err) {
   if (err) {
     throw err;
   }
-  mongoClient.connect('mongodb://localhost', function (err, client) {
+  mongoClient.connect('mongodb://127.0.0.1', function (err, client) {
     if (err) {
       throw err;
     }