+++ test/helpers/common.test.js
@@ -14,4 +14,11 @@ describe('modifiedPaths, bad update value which has circular reference field', (
 
     assert.throws(() => modifiedPaths(objA, 'path', null), /circular reference/);
   });
+
+  it('values with multiple references which are not circular should succeed', function() {
+    const objA = {};
+    const objB = { a: objA, b: objA };
+
+    modifiedPaths(objB, 'path', null);
+  });
 });