+++ test/model.test.js
@@ -6767,7 +6767,8 @@ describe('Model', function() {
         );
 
       });
-      xit('creates indexes only when they do not exist on the mongodb server (gh-12250)', async() => {
+
+      it('creates indexes only when they do not exist on the mongodb server (gh-12250)', async() => {
         const userSchema = new Schema({
           name: { type: String }
         }, { autoIndex: false });
@@ -6781,15 +6782,14 @@ describe('Model', function() {
         const createIndexSpy = sinon.spy(User.collection, 'createIndex');
         const listIndexesSpy = sinon.spy(User.collection, 'listIndexes');
 
-        // Act
         await User.syncIndexes();
+
         assert.equal(createIndexSpy.callCount, 1);
-        assert.equal(listIndexesSpy.callCount, 2);
+        assert.equal(listIndexesSpy.callCount, 1);
 
         await User.syncIndexes();
 
-        // Assert
-        assert.equal(listIndexesSpy.callCount, 4);
+        assert.equal(listIndexesSpy.callCount, 2);
         assert.equal(createIndexSpy.callCount, 1);
       });
     });