+++ types/middlewares.d.ts
@@ -2,6 +2,7 @@ declare module 'mongoose' {
 
   type MongooseDocumentMiddleware = 'validate' | 'save' | 'remove' | 'updateOne' | 'deleteOne' | 'init';
   type MongooseQueryMiddleware = 'count' | 'estimatedDocumentCount' | 'countDocuments' | 'deleteMany' | 'deleteOne' | 'distinct' | 'find' | 'findOne' | 'findOneAndDelete' | 'findOneAndRemove' | 'findOneAndReplace' | 'findOneAndUpdate' | 'remove' | 'replaceOne' | 'update' | 'updateOne' | 'updateMany';
+  type DocumentOrQueryMiddleware = 'updateOne' | 'deleteOne' | 'remove';
 
   type MiddlewareOptions = {
     /**